+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Revision history for salmon
 
+## 0.1.1.0 -- 2023-10-27
+
+* Added "game" sub-command
+    - Displays:
+        - Games won/lost/disconnected
+        - Win-rate/loss-rate/disconnect-rate
+        - Average wave reached
+* Internal reworks/code cleanup
+
 ## 0.1.0.0 -- 2023-08-21
 
 * First actual "release" version, with GitHub builds